backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
faiss-cpu
PyMuPDF
google-generativeai
pydub
cartesia
langchain
pydantic
numpy
motor
python-dotenv
